Took on the advice given to me.
got rid of the max-width from the tablet media query, got rid of the mobile query all together, eliminated duplicate code, changed the background colour to black to make the content pop out more with more contrast. made the box-shadow grey to still have contrast but not too much contrast. Made it less blurry to not draw attention to it